+++ src/main/java/com/bulain/jbpm4order/dao/GroupLoginMapper.java
@@ -6,10 +6,10 @@ import com.bulain.jbpm4order.pojo.GroupSearch;
 import com.bulain.jbpm4order.pojo.LoginSearch;
 
 public interface GroupLoginMapper extends BasicMapper<GroupLogin> {
-    int deleteGroupLoginByLoginId(Integer loginId);
+    int deleteGroupLoginByLoginId(Long loginId);
     int deleteGroupLoginByNotInLoginId(GroupSearch search);
 
-    int deleteGroupLoginByGroupId(Integer groupId);
+    int deleteGroupLoginByGroupId(Long groupId);
     int deleteGroupLoginByNotInGroupId(LoginSearch search);
 
     int deleteGroupLogin(LoginSearch search);
