+++ src/main/java/com/bulain/jbpm4order/dao/LoginMapper.java
@@ -7,8 +7,8 @@ import com.bulain.jbpm4order.model.Login;
 import com.bulain.jbpm4order.pojo.LoginSearch;
 
 public interface LoginMapper extends PagedMapper<Login, LoginSearch> {
-    Long countLoginByGroupId(Integer groupId);
-    List<Login> findLoginByGroupId(Integer groupId);
-    List<Login> findLoginByNoGroupId(Integer groupId);
+    Long countLoginByGroupId(Long groupId);
+    List<Login> findLoginByGroupId(Long groupId);
+    List<Login> findLoginByNoGroupId(Long groupId);
     List<Login> findLoginByLoginNames(String[] loginName);
 }
