+++ src/main/java/com/bulain/jbpm4order/dao/MailTemplateMapper.java
@@ -6,5 +6,5 @@ import com.bulain.jbpm4order.pojo.MailTemplateSearch;
 
 public interface MailTemplateMapper extends PagedMapper<MailTemplate, MailTemplateSearch> {
     int updateByPrimaryKeyWithBLOBs(MailTemplate record);
-    MailTemplate selectByPrimaryKeyWithoutBLOBs(Integer id);
+    MailTemplate selectByPrimaryKeyWithoutBLOBs(Long id);
 }
