+++ src/main/java/com/bulain/jbpm4order/dao/MasterMapper.java
@@ -6,7 +6,7 @@ import com.bulain.jbpm4order.pojo.Master;
 
 public interface MasterMapper {
     List<Master> selectList4Group();
-    Master selectMaster4Group(Integer id);
+    Master selectMaster4Group(Long id);
     List<Master> selectList4Person();
-    Master selectMaster4Person(Integer id);
+    Master selectMaster4Person(Long id);
 }
