+++ src/main/java/com/bulain/jbpm4order/model/Login.java
@@ -8,9 +8,9 @@ import com.bulain.common.model.Trackable;
 public class Login implements Trackable, Serializable {
     private static final long serialVersionUID = 5088470747456872995L;
 
-    private Integer id;
+    private Long id;
 
-    private Integer personId;
+    private Long personId;
 
     private String loginName;
 
@@ -28,19 +28,19 @@ public class Login implements Trackable, Serializable {
 
     private Date updatedAt;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getPersonId() {
+    public Long getPersonId() {
         return personId;
     }
 
-    public void setPersonId(Integer personId) {
+    public void setPersonId(Long personId) {
         this.personId = personId;
     }
