+++ src/main/java/com/bulain/jbpm4order/service/ContentService.java
@@ -5,5 +5,5 @@ import com.bulain.jbpm4order.model.Content;
 import com.bulain.jbpm4order.pojo.ContentSearch;
 
 public interface ContentService extends PagedService<Content, ContentSearch> {
-    Content getWithoutBLOBs(Integer id);
+    Content getWithoutBLOBs(Long id);
 }
