+++ src/main/java/com/bulain/jbpm4order/service/MasterService.java
@@ -5,8 +5,8 @@ import java.util.List;
 import com.bulain.jbpm4order.pojo.Master;
 
 public interface MasterService {
-    String getValue4Group(Integer id);
+    String getValue4Group(Long id);
     List<Master> findMaster4Group();
-    String getValue4Person(Integer id);
+    String getValue4Person(Long id);
     List<Master> findMaster4Person();
 }
